remove \flushbottom
\flushbottom makes all pages have exactly the height specified in the
geometry settings.  this is impossible when the page height is not a
multiple of the height of one line of text (as is curently the case)
and there is no stretchable space.  it is also impossible on the first
page when the page height minus the titlebox height is not a multiple
of the line height.  this leads to the two columns ending at different
heights when latex can make one column fit the page height, but not
the other.

this commit removes \flushbottom.  now the two columns always match,
and the underfull vbox errors are gone, but the page heights of the
different pages of the paper are not always the same.  I think that
is probably fine.  ideally, \textheight-\top and \titlebox should
both be multiples of \baselineskip.

Fixes #11