Enable pdflatex+hyperref to generate ToC-metadata
The style file was configured so that LaTeX would not to generate a table of contents. It was stated that this is “never going to [be] need[ed].” This configuration has been part of the ACL style files since, at least, [ACL 2000](https://cse.hkust.edu.hk/acl2000/Latex/aclsub.sty).

I do not agree that the table of contents is not needed. While it may be true that a table of contents will never be printed on the first page of an ACL-paper, these lines also prevent `hyperref` from generating the toc-metadata that PDF-viewers use to show an electronic table of content for navigating the PDF, a feature that I find immensely useful.

This pull commit deletes these three lines from the style file and re-enables the generation of ToC metadata