Do not redefine \par in preamble.
The lineno package's \linenumbers redefine \par which causes a bug in
babel's luatex code where a \par expands to some lineno macro which do
not constitute valid lua code resulting in an error.
This commit delay \linenumbers's redefinition of \par to the start of
the document to avoid the error.